refactor(repo): modernize docs architecture, backend services, and frontend tooling
Summary:
- Replaced legacy `.ai/knowledge/` and Zephyr PKF files with Token Atlas Lite specifications (`.ai/INDEX.md`, `MEMORY.md`, `DECISIONS.md`, `DEPENDENCIES.md`, `GLOSSARY.md`).
- Migrated Python project setup and container builds to `uv` (`pyproject.toml`, `uv.lock`, `.python-version`, `Dockerfile`, CI workflows).
- Refactored core backend modules (`auth`, `items`, `sales`, `metal_rates`, `dashboard`) to add invoice PDF generation, refined pricing calculations, and improved security auditability.
- Upgraded frontend setup to flat ESLint configuration, Vitest test runner, TanStack Query keys management, and typed storage utilities.
- Added comprehensive unit and integration test coverage across backend (`tests/`) and frontend (`frontend/src/`).

Rationale:
- Reduces AI context overhead and simplifies repo metadata maintenance using Token Atlas Lite.
- Accelerates backend dependency installation and deterministic build reproducibility using `uv`.
- Improves domain calculation precision (metal rates & item pricing) and user authentication resilience.
- Enhances frontend developer experience, code reliability, and automated testing capabilities.

Tests:
- Backend: `uv lock --check`, `uv run pytest`
- Frontend: `npm run lint`, `npm run typecheck`, `npm test`

+++ app/modules/metal_rates/schemas.py
@@ -1,7 +1,9 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class MetalRateCreate(BaseModel):
     metal: str
-    purity: float
-    rate_per_gram: float
+    purity: Decimal
+    rate_per_gram: Decimal
